backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
google-genai==0.2.2
pydantic==2.9.2
python-dotenv==1.0.1
